requirements/development.txt
-r base.txt

# [drf spectacular dependencies]
attrs==25.2.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
uritemplate==4.1.1

# [debug tools dependencies]
django-debug-toolbar==5.0.1

# [pytest dependencies]
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.10.0
pytest-mock==3.14.0
